1、优化SensorSetting界面传感器编号内容显示为出厂编号 2、发布V2.5版本

+++ THTS/DeviceCenter/DeviceCenterViewModel.cs
@@ -156,6 +156,11 @@ namespace THTS.DeviceCenter
         /// </summary>
         private void Delete()
         {
+            if(DeviceSelected == null || DeviceSelected.SensorId == null)
+            {
+                return;
+            }
+
             bool delete = DataAccess.EntityDAO.DeviceDAO.Delete(DeviceSelected);
             if (delete)
             {
